Added depency check in: restockOrderDBU and testDescriptorDBU

+++ code/server/database_utilities/restockOrderDBU.js
@@ -12,92 +12,92 @@ class RestockOrderDBU {
     // - db (Database)
 
     // constructor
-    constructor(dbname, db=undefined) {
+    constructor(dbname, db = undefined) {
         if (!db) {
             this.db = new sqlite.Database(dbname, (err) => {
                 if (err) throw err;
             });
         } else {
             this.db = db;
-        } 
+        }
     }
     close() {
         this.db.close();
     }
 
-// get RestockOrder(s) from the RESTOCK-ORDERS table and return it/them as a RestockOrder object
-    loadRestockOrder(id=undefined, state=undefined, returnItems=undefined) { //returnItems is a flag 
-            const sqlInfo = {sql: undefined, values: undefined};
+    // get RestockOrder(s) from the RESTOCK-ORDERS table and return it/them as a RestockOrder object
+    loadRestockOrder(id = undefined, state = undefined, returnItems = undefined) { //returnItems is a flag 
+        const sqlInfo = { sql: undefined, values: undefined };
 
-            if(id) {
-                if(!returnItems) {
-                    const sqlId = 'SELECT * FROM "restock-orders" WHERE id = ?';
-                    sqlInfo.sql = sqlId;
-                    sqlInfo.values = [id];
+        if (id) {
+            if (!returnItems) {
+                const sqlId = 'SELECT * FROM "restock-orders" WHERE id = ?';
+                sqlInfo.sql = sqlId;
+                sqlInfo.values = [id];
+            }
+            else {
+                // return an array
+                if (!this.#checkState(orderId, 'COMPLETEDRETURN')) {
+                    throw (new Error("Incorrect order's status. Operation aborted", 13));
                 }
-                else {
-                    // return an array
-                    if(!this.#checkState(orderId, 'COMPLETEDRETURN')) {
-                        throw(new Error("Incorrect order's status. Operation aborted", 13));
-                    }
-                    const sqlReturn = 'SELECT S.SKUid as skuId, S.RFID as rfid FROM "sku-items" S, "sku-items-rko" R, "test-results" Tr WHERE R.orderId=? AND s.id = r.skuItemId AND R.skuItemId=Tr.SKUitemId AND Tr.result= Fail'
-                    sqlInfo.sql = sqlReturn;
-                    sqlInfo.values = [id]
+                const sqlReturn = 'SELECT S.SKUid as skuId, S.RFID as rfid FROM "sku-items" S, "sku-items-rko" R, "test-results" Tr WHERE R.orderId=? AND s.id = r.skuItemId AND R.skuItemId=Tr.SKUitemId AND Tr.result= Fail'
+                sqlInfo.sql = sqlReturn;
+                sqlInfo.values = [id]
 
-                    return new Promise((resolve, reject) => {
-                        this.db.all(sqlInfo.sql, sqlInfo.values, (err, rows) => {
-                            if (err) {
-                                reject(err);
-                                return;
-                            }
-                            const skuItemArray = rows.map((si) => {
-                                const sia = {SKUid: rows.skuId, rfid: rows.rfid};
-                                return sia;
-                            });
-                            Promise.all(skuItemArray).then((skuItemArray) => resolve(skuItemArray));
-                        });
-                    });
-                    
-                }
-            }
-            else if(state) {
-                const sqlState = 'SELECT * FROM "restock-orders" WHERE state = ?';
-                sqlInfo.sql = sqlState;
-                sqlInfo.values = [state];
-            }
-            else{
-                const sqlNoInfo = 'SELECT * FROM "restock-orders"';
-                sqlInfo.sql = sqlNoInfo;
-                sqlInfo.values = [];
-            }
-            return new Promise((resolve, reject) => {
-                this.db.all(sqlInfo.sql, sqlInfo.values, (err, rows) => {
-                    if (err) {
-                        reject(err);
-                        return;
-                    }
-                    const orders = rows.map(async (o) => {
-                        const order = new RestockOrder(o.id, o.issueDate, o.state, o.supplierId, o.transportNote);
-                        if(o.state !== 'DELIVERY' && o.state !== 'ISSUED') {
-                            const products = await this.#getProducts(o.id);
-                            const skuItems = await this.#getSkuItems(o.id);
-                            order.setProducts(products);
-                            order.setSkuItems(skuItems);
+                return new Promise((resolve, reject) => {
+                    this.db.all(sqlInfo.sql, sqlInfo.values, (err, rows) => {
+                        if (err) {
+                            reject(err);
+                            return;
                         }
-                        return order;
+                        const skuItemArray = rows.map((si) => {
+                            const sia = { SKUid: rows.skuId, rfid: rows.rfid };
+                            return sia;
+                        });
+                        Promise.all(skuItemArray).then((skuItemArray) => resolve(skuItemArray));
                     });
-                    Promise.all(orders).then((orders) => resolve(orders));
                 });
-            });
-        }
 
-// insert a new RestockOrder inside the RESTOCK-ORDERS table
+            }
+        }
+        else if (state) {
+            const sqlState = 'SELECT * FROM "restock-orders" WHERE state = ?';
+            sqlInfo.sql = sqlState;
+            sqlInfo.values = [state];
+        }
+        else {
+            const sqlNoInfo = 'SELECT * FROM "restock-orders"';
+            sqlInfo.sql = sqlNoInfo;
+            sqlInfo.values = [];
+        }
+        return new Promise((resolve, reject) => {
+            this.db.all(sqlInfo.sql, sqlInfo.values, (err, rows) => {
+                if (err) {
+                    reject(err);
+                    return;
+                }
+                const orders = rows.map(async (o) => {
+                    const order = new RestockOrder(o.id, o.issueDate, o.state, o.supplierId, o.transportNote);
+                    if (o.state !== 'DELIVERY' && o.state !== 'ISSUED') {
+                        const products = await this.#getProducts(o.id);
+                        const skuItems = await this.#getSkuItems(o.id);
+                        order.setProducts(products);
+                        order.setSkuItems(skuItems);
+                    }
+                    return order;
+                });
+                Promise.all(orders).then((orders) => resolve(orders));
+            });
+        });
+    }
+
+    // insert a new RestockOrder inside the RESTOCK-ORDERS table
     async insertRestockOrder(issueDate, products, supplierId, transportNote) {
         // check if supplier exist
         const isSupplier = await this.#checkSupplier(supplierId);
         if (!isSupplier)
-            throw(new Error("Supplier does not exist. Operation aborted.", 6));
-    
+            throw (new Error("Supplier does not exist. Operation aborted.", 6));
+
         const orderId = await this.#insertOrder(issueDate, supplierId, transportNote);
         const promises = products.map((p) => new Promise(async (resolve, reject) => {
             const insert = 'INSERT INTO "products-rko" (orderId, skuId, description, price, quantity) VALUES (?,?,?,?,?)';
@@ -111,17 +111,17 @@ class RestockOrderDBU {
         return Promise.all(promises);
     }
 
-// update a selected RestockOrder in the RESTOCK-ORDERS table. Return number of rows modified
+    // update a selected RestockOrder in the RESTOCK-ORDERS table. Return number of rows modified
     async updateRestockOrder(orderId, newState = undefined, newTransportNote = undefined, skuItems = undefined) {
-        const sqlInfo = {sql: undefined, values: undefined}
+        const sqlInfo = { sql: undefined, values: undefined }
         // check if newState exist
-        if(newState) {
+        if (newState) {
             sqlInfo.sql = 'UPDATE "restock-orders" SET state = ? WHERE id = ?';
             sqlInfo.values = [newState, orderId];
             //update state
             return new Promise((resolve, reject) => {
                 this.db.run(sqlInfo.sql, sqlInfo.values, function (err) {
-                    if(err) {
+                    if (err) {
                         reject(err);
                         return;
                     }
@@ -132,43 +132,43 @@ class RestockOrderDBU {
             });
         }
         // check if SKUitem(s) are passed
-        else if(skuItems) {
-            if(!this.#checkState(orderId, 'DELIVERED'))
-                throw(new Error("Incorrect order's status. Operation aborted", 13));
+        else if (skuItems) {
+            if (!this.#checkState(orderId, 'DELIVERED'))
+                throw (new Error("Incorrect order's status. Operation aborted", 13));
 
             const ids = [];
-            for(ski of skuItems){
+            for (ski of skuItems) {
                 // check if SKUitem exists
                 const isSKUitem = await this.#checkSKUitem(ski);
-                
+
                 if (!isSKUitem)
-                    throw(new Error("SKUitem does not exist. Operation aborted.", 9));
+                    throw (new Error("SKUitem does not exist. Operation aborted.", 9));
                 // save SkuItemId and SKUid
                 else
-                    ids.push({siId: isSKUitem,skId: ski.SKUId});
+                    ids.push({ siId: isSKUitem, skId: ski.SKUId });
             }
-              const promises = ids.map((i) => new Promise( async (resolve, reject) => {
-                  const addItem = 'INSERT INTO "sku-items-rko" (orderId, skuItemId, skuId) VALUES (?,?,?)';
-                  this.db.run(addItem, [orderId, i.siId, i.skId], function (err) {
-                      if(err) {
-                          reject(err);
-                          return;
-                      } else resolve('Done');
-                  });
-              }));
-              await Promise.all(promises);
+            const promises = ids.map((i) => new Promise(async (resolve, reject) => {
+                const addItem = 'INSERT INTO "sku-items-rko" (orderId, skuItemId, skuId) VALUES (?,?,?)';
+                this.db.run(addItem, [orderId, i.siId, i.skId], function (err) {
+                    if (err) {
+                        reject(err);
+                        return;
+                    } else resolve('Done');
+                });
+            }));
+            await Promise.all(promises);
         }
         // check if newTransportNote exists
-        if(newTransportNote) {
-            if(!this.#checkState(orderId, 'DELIVERY'))
-                throw(new Error("Incorrect order's status. Operation aborted", 13));
+        if (newTransportNote) {
+            if (!this.#checkState(orderId, 'DELIVERY'))
+                throw (new Error("Incorrect order's status. Operation aborted", 13));
 
             sqlInfo.sql = 'UPDATE "restock-orders" SET transportNote = ? WHERE id = ?';
             sqlInfo.values = [newTransportNote, orderId];
             // update TransportNote
             return new Promise((resolve, reject) => {
                 this.db.run(sqlInfo.sql, sqlInfo.values, function (err) {
-                    if(err) {
+                    if (err) {
                         reject(err);
                         return;
                     }
@@ -180,42 +180,47 @@ class RestockOrderDBU {
         }
     }
 
-// delete one or more RestockOrder from the RESTOCK-ORDERS table given different input. Return number of rows modified
+    // delete one or more RestockOrder from the RESTOCK-ORDERS table given different input. Return number of rows modified
     async deleteRestockOrder(orderId) {
-            const promises = [];
-            // delete from restock orders
-            promises.push(new Promise((resolve, reject) => {
-                const sqlDelete = 'DELETE FROM "restock-orders" WHERE id=?';
-                this.db.run(sqlDelete, [orderId], function (err) {
-                    if (err) {
-                        reject(err);
-                        return;
-                    } else resolve(this.changes);
-                });
-            }));
-            // delete from products
-            promises.push(new Promise((resolve, reject) => {
-                const sqlDelete = 'DELETE FROM "products-rko" WHERE orderId=?';
-                this.db.run(sqlDelete, [orderId], function (err) {
-                    if (err) {
-                        reject(err);
-                        return;
-                    } else resolve(this.changes);
-                });
-            }));
-            // delete from skuItem Rko
-            promises.push(new Promise((resolve, reject) => {
-                const sqlDelete = 'DELETE FROM "sku-items-rko" WHERE orderId=?';
-                this.db.run(sqlDelete, [orderId], function (err) {
-                    if (err) {
-                        reject(err);
-                        return;
-                    } else resolve(this.changes);
-                });
-            }));
-            return Promise.all(promises);
+        const dependency = await this.#checkDependency(id);
+        if (dependency) {
+            // if there is at least 1 dependency
+            throw (new Error("Dependency detected. Delete aborted.", 14));
         }
-    
+        const promises = [];
+        // delete from restock orders
+        promises.push(new Promise((resolve, reject) => {
+            const sqlDelete = 'DELETE FROM "restock-orders" WHERE id=?';
+            this.db.run(sqlDelete, [orderId], function (err) {
+                if (err) {
+                    reject(err);
+                    return;
+                } else resolve(this.changes);
+            });
+        }));
+        // delete from products
+        promises.push(new Promise((resolve, reject) => {
+            const sqlDelete = 'DELETE FROM "products-rko" WHERE orderId=?';
+            this.db.run(sqlDelete, [orderId], function (err) {
+                if (err) {
+                    reject(err);
+                    return;
+                } else resolve(this.changes);
+            });
+        }));
+        // delete from skuItem Rko
+        promises.push(new Promise((resolve, reject) => {
+            const sqlDelete = 'DELETE FROM "sku-items-rko" WHERE orderId=?';
+            this.db.run(sqlDelete, [orderId], function (err) {
+                if (err) {
+                    reject(err);
+                    return;
+                } else resolve(this.changes);
+            });
+        }));
+        return Promise.all(promises);
+    }
+
     // private method to get products for a given orderId 
     #getProducts(id) {
         return new Promise((resolve, reject) => {
@@ -272,17 +277,17 @@ class RestockOrderDBU {
             });
         });
     }
-    
+
     //private method to check whether the state of the Order is DELIVERED
     #checkState(orderId, status) {
         const sql = 'SELECT * FROM "restock-orders" WHERE id = ? AND state = ?';
         return new Promise((resolve, reject) => {
             this.db.get(sql, [orderId, status], (err, row) => {
-                if(err) {
+                if (err) {
                     reject(err);
                     return;
                 }
-                resolve(row ? true: false);
+                resolve(row ? true : false);
             });
         });
     }
@@ -292,11 +297,28 @@ class RestockOrderDBU {
         const sql = 'SELECT id FROM "SKU-ITEMS" WHERE RFID=? AND SKUid = ?'
         return new Promise((resolve, reject) => {
             this.db.get(sql, [skuItem.RFID, skuItem.SKUId], (err, row) => {
-                if(err) {
+                if (err) {
                     reject(err);
                     return;
                 }
                 resolve(row ? row.id : false);
+            });
+        });
+    }
+
+    #checkDependency(id) {
+        // return-order check
+        return new Promise((resolve, reject) => {
+            const returnOrder = 'SELECT id FROM "return-order" WHERE restockOrderId=?';
+            this.db.all(returnOrder, [id], (err, rows) => {
+                if (err) {
+                    reject(err);
+                    return;
+                }
+                if (!rows || rows.length == 0)
+                    resolve(true);
+                else resolve(false);
+                return;
             });
         });
     }
